move condtional check from service to controller

+++ src/main/java/com/example/demo/controller/StudentController.java
@@ -1,15 +1,17 @@
 package com.example.demo.controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import com.example.demo.entity.Student;
 import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.example.demo.entity.Student;
 import com.example.demo.serviceimp.studentService;
 
 @RestController
@@ -38,7 +40,7 @@ public class StudentController {
         return studentService.getStudentById(id);
     }
 
-    @PutMapping("/{id}")
+    @PutMapping
     public Student updateStudent(@RequestBody Student studentDetails) {
         return studentService.updateStudent(studentDetails);
     }
@@ -47,7 +49,10 @@ public class StudentController {
 
     public void deletestudent(@PathVariable Long id) {
 
-        studentService.deleteStudent(id);
+       if(studentService.getStudentById(id).getId().equals(id)) {
+
+           studentService.deleteStudent(id);
+       }
 
     }
